src/requirements.txt
# Python 3.12 - Latest Stable Versions
# LangChain and LangGraph 1.0.x

# Core LangChain
langgraph==1.0.1
langchain==1.0.2
langchain-community==0.4

# AWS Integration
langchain-aws==1.0.0

# Google Gemini
langchain-google-genai

# Database
trino[sqlalchemy]

# FastAPI
fastapi
uvicorn
websockets

# Utilities
python-dotenv

# Other
